requirements.txt
requests
pytesseract
pillow
langdetect
python-dotenv
